requirements.txt
appdirs==1.4.3
Django==1.11
django-ckeditor==5.2.1
django-solo==1.1.2
easy-thumbnails==2.4.1
olefile==0.44
packaging==16.8
Pillow==4.1.0
psycopg2==2.7.1
pyparsing==2.2.0
pytz==2017.2
six==1.10.0
